+++ ui/CHANGES.md
@@ -3,6 +3,7 @@
 - **Added**: `VisTextField#clearText()`
 - **Added**: `FloatingGroup`
 - **Added**: `VisWindow#isKeepWithinParent` and `VisWindow#setKeepWithinParent` 
+- **Added**: constructor `VisImage (String drawableName)`
 - **Fixed**: `OptionDialog#set(...)ButtonText` now updates dialog size
 - **Fixed**: [#131](https://github.com/kotcrab/vis-editor/issues/131) - fixed issue when copying numbers between `VisTextField`s with `FloatDigitsOnlyFilter` decimal point was lost
 - **Fixed**: `ListView#AbstractListAdapter` error on GWT